Extended searching for customer users
Znuny does not allow to search for customer users using first name
and surname separated with space (natural way of searching) -
if one search for "john black", fields specified
in CustomerUserSearchFields will be searched for this whole phrase
individually and customer user with firstname = John
and surname = Black won't match. To be able to search using
firstname and surname, one need to use "john+black" syntax which
is not obvious.

This mod introduces new parameter for CustomerUser searching maps
(to be configured in Config.pm):

    CustomerUserSearchExtended => 0,   # 0 - normal search (space is part of search phrase), 1 - extended search (space works like +)

When set to 1, Znuny treats whitespaces in searched string same
like +; when searching for "john black" all searched fields will
be searched separate for john and black.

Author-Change-Id: IB#1055633