Added optional Feature "DontSkipMe"
This makes it possible to put several TransitionActions of the type TicketCreate into one transition and to process them depending on e.g. checkboxes. Example: several dynamic fields of the type Checkbox - each checkbox is to cause a split ticket to be created. Currently, wild transitions have to be built to solve this problem. With this change you can simply paste the value of the checkbox <OTRS_TICKET_DynamicField_Checkbox1> to the config parameter DontSkipMe and get the desired result.