Fixed error "Can't sign: unable to write 'random state'" if sending emails signed with S/MIME. – OTRS bug #14522 (#72).
OpenSSL generates a .rnd ito sign emails with S/MIME. This causes an error "Can't sign: unable to write 'random state'" in the system log, because the .rnd file has only user r/w rights, but not group rights. If you send a email yourself (e. g. answering a customer), the email is send by the apache process, so the .rnd file is owned by the apache user. (If you set the group rights manually, this will be undone by the system).