List all configured calendars in support data collector. (#185)
One can define more than just the default 9 calendars. In AdminSLA and AdminQueue this is already implemented, but not in the support data collector. This PR fixes that.

The support data list the calendars their configured time zones. If there are more than 9 calendars in the system configuration, the support data just list the calendars 1..9. After applying this patch, all the configured calendars are listed.

The calendar name is checked to check if the calendar is configured. Otherwise the support data would list 50 calendars (or whatever is configured as the maximum number of calendars) no matter if the calendars exist in the system configuration or not.